+++ .gitignore
@@ -157,4 +157,15 @@ cython_debug/
 #  be found at https://github.com/github/gitignore/blob/main/Global/JetBrains.gitignore
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
-.idea/
+idea/
+src/lab_3/Data_preparation
+src/lab_3/Data_validation
+src/lab_3/Model_creation
+src/lab_3/Model_evaluation
+src/lab_3/Model_save
+src/tests/
+*.joblib
+.telemetry
+wandb/
+stats.json
+session_store.db
