--- src/Kedro/pipelines/model_evaluation/__init__.py
@@ -1,10 +1,0 @@
-"""
-This is a boilerplate pipeline 'model_evaluation'
-generated using Kedro 0.19.3
-"""
-
-from .pipeline import create_pipeline
-
-__all__ = ["create_pipeline"]
-
-__version__ = "0.1"
