+++ src/lab_3/Kedro/pipeline_registry.py
@@ -1,0 +1,17 @@
+"""Project pipelines."""
+from typing import Dict
+
+from kedro.framework.project import find_pipelines
+from kedro.pipeline import Pipeline
+
+
+
+def register_pipelines() -> Dict[str, Pipeline]:
+    """Register the project's pipelines.
+
+    Returns:
+        A mapping from pipeline names to ``Pipeline`` objects.
+    """
+    pipelines = find_pipelines()
+    pipelines["__default__"] = sum(pipelines.values())
+    return pipelines
