+++ src/lab_3/Kedro/pipelines/data_processing/__init__.py
@@ -1,0 +1,1 @@
+from .pipeline import create_pipeline  # NOQA
